requirements.txt
asgiref==3.8.1
cloudinary==1.25.0
Django==5.0.7
django-cloudinary-storage==0.3.0
django-filter==24.2
djangorestframework==3.15.2
pillow==10.4.0
pytz==2024.1
sqlparse==0.5.1
